requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.4
boto3==1.9.41
botocore==1.12.41
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cloudinary==1.15.0
cryptography==2.4.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.36.0
django-appconf==1.0.2
django-bootstrap-forms==0.1
django-bootstrap4==0.0.7
django-celery-beat==1.5.0
django-celery-email==3.0.0
django-ckeditor==5.7.1
django-cloudinary-storage==0.2.3
django-cors-headers==3.0.2
django-countries==5.3.2
django-crispy-forms==1.7.0
django-filter==2.0.0
django-filters==0.2.1
django-forms-bootstrap==3.1.0
django-formtools==2.1
django-heroku==0.3.1
django-invitations==1.9.2
django-js-asset==1.2.2
django-mailgun==0.9.1
django-multiselectfield==0.1.8
django-paypal==0.5.0
django-phonenumber-field==2.1.0
django-rest-auth==0.9.5
django-separatedvaluesfield==0.6.0
django-storages==1.7.1
django-taggit==0.24.0
django-timezone-field==4.0
djangorestframework==3.9.4
djangorestframework-api-key==1.4.1
djangorestframework-jwt==1.11.0
docutils==0.14
enum34==1.1.6
Fabric3==1.14.post1
gunicorn==19.7.1
idna==2.7
importlib-metadata==1.2.0
ipaddress==1.0.17
jeepney==0.4
jmespath==0.9.3
jsonpickle==1.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
Markdown==3.1
martor==1.4.0
mock==3.0.5
more-itertools==8.0.2
oauthlib==2.1.0
paramiko==2.4.2
phonenumbers==8.10.0
Pillow==5.2.0
psycopg2==2.7.6.1
psycopg2-binary==2.8.4
pyasn1==0.4.4
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
python-crontab==2.4.0
python-dateutil==2.7.5
python-decouple==3.1
python-digitalocean==1.14.0
python3-openid==3.1.0
pytz==2018.4
pywin32-ctypes==0.2.0
pyxdg==0.25
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.3.0
urllib3==1.23
vine==1.1.4
whitenoise==3.3.1
zipp==0.6.0
